centralise NQueens Node implementation
