+++ baselines/nqueens/nqueens_gpu_cuda.cu
@@ -10,29 +10,11 @@
 #include <time.h>
 #include "cuda_runtime.h"
 
+#include "lib/NQueens_node.h"
+
 #define BLOCK_SIZE 512
 
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
-
-/*******************************************************************************
-Implementation of N-Queens Nodes.
-*******************************************************************************/
-
-#define MAX_QUEENS 20
-
-typedef struct
-{
-  uint8_t depth;
-  uint8_t board[MAX_QUEENS];
-} Node;
-
-void initRoot(Node* root, const int N)
-{
-  root->depth = 0;
-  for (uint8_t i = 0; i < N; i++) {
-    root->board[i] = i;
-  }
-}
 
 /*******************************************************************************
 Implementation of a dynamic-sized single pool data structure.
